Created utility file and added similar functions in said utility file
