+++ enc_client.c
@@ -6,6 +6,7 @@
 #include <sys/socket.h> // send(),recv()
 #include <netdb.h>      // gethostbyname()
 #include <fcntl.h>
+#include "utility.h"
 
 #define BUFF_SIZE 70000
 
@@ -15,104 +16,6 @@
 * 2. Prompt the user for input and send that input as a message to the server.
 * 3. Print the message received from the server and exit the program.
 */
-
-// Checks to see if plaintext has any incorrect letters 1 = true, 0 = false.
-int isIncorrect(char *string)
-{
-  int i = 0;
-  for (i; i < strlen(string); i++)
-  {
-    char c = string[i];
-    int isLetter = 0;
-    if (c >= 65 && c <= 90)
-      isLetter = 1;
-    int isSpace = 0;
-    if (c == 32)
-      isSpace = 1;
-    int isCrNl = 0;
-    if (c == '\n')
-      isCrNl = 1;
-    // If the character is not within acceptable bounds then return
-    // that string is incorrect
-    if (isLetter == 0 && isSpace == 0 && isCrNl == 0)
-      return 1;
-  }
-  return 0;
-}
-
-// Error function used for reporting issues
-void error(const char *msg)
-{
-  perror(msg);
-  exit(0);
-}
-
-// Set up the address struct
-void setupAddressStruct(struct sockaddr_in *address, int portNumber)
-{
-  // Clear out the address struct
-  memset((char *)address, '\0', sizeof(*address));
-
-  // The address should be network capable
-  address->sin_family = AF_INET;
-  // Store the port number
-  address->sin_port = htons(portNumber);
-
-  // Host name should be localhost
-  struct hostent *hostInfo = gethostbyname("localhost");
-  if (hostInfo == NULL)
-  {
-    fprintf(stderr, "CLIENT: ERROR, no such host\n");
-    exit(0);
-  }
-  // Copy the first IP address from the DNS entry to sin_addr.s_addr
-  memcpy((char *)&address->sin_addr.s_addr, hostInfo->h_addr_list[0], hostInfo->h_length);
-}
-
-// Send all method
-// https://stackoverflow.com/a/14184228/6889483
-int send_all(int socket, char *buffer, size_t length, int flags)
-{
-    // Use charsWritten to keep track of how many characters are written
-    ssize_t charsWritten;
-    // Use buffIdx to advance buffer pointer
-    char* buffIdx = buffer;
-    // Keep on looping until length reachs 0
-    while (length > 0)
-    {
-        charsWritten = send(socket, buffIdx, length, flags);
-        if (charsWritten <= 0)
-            return -1;
-        buffIdx += charsWritten;
-        length -= charsWritten;
-    }
-    return 0;
-}
-
-// Receive all method (similar to send_all method)
-// Uses termination character '@' to stop receving message
-// https://stackoverflow.com/a/16256724/6889483
-int recv_all(int socket, char *buffer)
-{
-    memset(buffer, '\0', BUFF_SIZE);
-    ssize_t charsWritten;
-    char* buffIdx = buffer;
-    int length = BUFF_SIZE;
-    // Will stop looping until termination character is in buffer
-    while (strchr(buffer, '@') == NULL)
-    {
-        // Keep on looping until buffIdx is equal to greater than totalSize of message
-        charsWritten = recv(socket, buffIdx, length, 0);
-        if (charsWritten <= 0)
-            return -1;
-        buffIdx += charsWritten;
-        length -= charsWritten;
-    }
-    // Replace the termination character with a null character in buffer before returning
-    buffer[charsWritten - 1] = '\0';
-    return 0;
-}
-
 
 int main(int argc, char *argv[])
 {
@@ -174,23 +77,23 @@ int main(int argc, char *argv[])
   socketFD = socket(AF_INET, SOCK_STREAM, 0);
   if (socketFD < 0)
   {
-    error("CLIENT: ERROR opening socket");
+    clientError("CLIENT: ERROR opening socket");
   }
 
   // Set up the server address struct
-  setupAddressStruct(&serverAddress, atoi(argv[3]));
+  setupAddressStructClient(&serverAddress, atoi(argv[3]));
 
   // Connect to server
   if (connect(socketFD, (struct sockaddr *)&serverAddress, sizeof(serverAddress)) < 0)
   {
     printf("CLIENT: ERROR connecting");
-    error("CLIENT: ERROR connecting");
+    clientError("CLIENT: ERROR connecting");
   }
   // Let server know that you are an encryption client
   int enc_msg = send_all(socketFD, "ENC_CLIENT@", 11, 0);
   // Receive server message that it's okay to continue
   int server_recv = recv_all(socketFD, buffer);
-  if (server_recv < 0) error("Client ERROR reading from socket for cipher");
+  if (server_recv < 0) clientError("Client ERROR reading from socket for cipher");
   if (strcmp(buffer, "OK") != 0)
   {
       fprintf(stderr, "ENC_CLIENT tried to connect to incorrect server!\n");
@@ -202,24 +105,20 @@ int main(int argc, char *argv[])
 
       // Receive server message for plaintext
       int first_recv = recv_all(socketFD, buffer);
-      if (first_recv < 0) error("ERROR reading from socket for plaintext");
-      //printf("Received from server: %s\n", buffer);
-      //fflush(stdout);
+      if (first_recv < 0) clientError("ERROR reading from socket for plaintext");
 
       int sec_msg = send_all(socketFD, key_buff, totalSize, 0);
 
       // Receive server message for key
       int sec_recv = recv_all(socketFD, buffer);
-      if (sec_recv < 0) error("ERROR reading from socket for key");
-      //printf("Received from server: %s\n", buffer);
-      //fflush(stdout);
+      if (sec_recv < 0) clientError("ERROR reading from socket for key");
 
       // Send wait message
       int send_wait = send_all(socketFD, "Waiting for cipher text@", 24, 0);
 
       // Receive server message for cipher
       int cipher_recv = recv_all(socketFD, buffer);
-      if (cipher_recv < 0) error("ERROR reading from socket for cipher");
+      if (cipher_recv < 0) clientError("ERROR reading from socket for cipher");
       printf("%s\n", buffer);
       fflush(stdout);
   }
